requirements.txt
fastapi[standard]
# uvicorn
httpx
python-dotenv
passlib
pymongo[srv]
openai